pix2code_project/requirements.txt
tensorflow==2.14.0
tensorflow-addons==0.23.0
keras-cv==0.8.2
numpy>=1.19.2
Pillow>=8.0.0
matplotlib>=3.3.2
pandas>=1.1.3
scikit-learn>=0.23.2
scikit-image>=0.17.2
opencv-python>=4.4.0
nltk>=3.5
pyyaml>=5.3.1
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pillow>=9.5.0
numpy>=1.24.0
tensorboard>=2.13.0
pyyaml>=6.0
timm>=0.9.0
torch-geometric>=2.3.0
nltk>=3.8.1
fiftyone>=0.20.0 